convert time to ghamari .

+++ app/src/main/java/com/app/weather/model/forecast/ListItem.kt
@@ -8,7 +8,7 @@ import com.google.gson.annotations.SerializedName
 
 data class ListItem(
     val clouds: Clouds?,
-    val dt: Int?,
+    val dt: Long?,
     @SerializedName("dt_txt")
     val dtTxt: String?,
     val main: Main?,
